requirements.txt
alembic==1.16.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
stripe==12.3.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
